+++ net/source/hcl/simulation/waveformFormats/MemoryTrace.h
@@ -23,7 +23,7 @@ struct MemoryTrace {
     };
 
     struct SignalChange {
-        unsigned sigIdx;
+        size_t sigIdx;
         size_t dataOffset;
     };
